moved Enzo-specific code related to refinement criteria, prolongation, and restriction into Enzo/amr
