+++ src/Enzo/CMakeLists.txt
@@ -47,6 +47,7 @@ target_include_directories(enzo
 target_link_options(enzo PRIVATE ${Cello_TARGET_LINK_OPTIONS})
 
 # add source files from various subdirectories to the enzo target
+add_subdirectory(amr)
 add_subdirectory(gravity)
 add_subdirectory(hydro-mhd) # add files in hydro-mhd to enzo
 add_subdirectory(initial) # add files in initial to enzo
